# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: INDOPACOM IRONVEIN (Resilient Operational Network for Vital Expeditionary Inventory Nodes) API for medical logistics decision support.
servers:
  - url: /api
...
paths:
...
  /predictive/forecast:
    post:
      operationId: forecastDemand
      tags: [predictive]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForecastInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForecastResult" }
  /predictive/recommendations:
    get:
      operationId: getRecommendations
      tags: [predictive]
      parameters:
        - in: query
          name: nodeId
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Recommendation" }
  /predictive/recommendations/{recommendationId}/promote:
    post:
      operationId: promoteRecommendationToOrder
      tags: [predictive]
      parameters:
        - in: path
          name: recommendationId
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PromoteRecommendationOverrides" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
components:
  schemas:
...
    Order:
      type: object
      required: [id, orderNumber, status, fromNodeId, toNodeId, supplierId, itemId, quantity, priority, etaDays, totalCost, createdAt]
      properties:
        id: { type: string }
        orderNumber: { type: string }
        status: { type: string }
        fromNodeId: { type: string }
        fromNodeName: { type: ["string", "null"] }
        toNodeId: { type: string }
        toNodeName: { type: ["string", "null"] }
        supplierId: { type: ["string", "null"] }
        supplierName: { type: ["string", "null"] }
        itemId: { type: string }
        itemName: { type: string }
        unit: { type: ["string", "null"] }
        quantity: { type: number }
        priority: { type: string }
        etaDays: { type: number }
        requestedDeliveryAt: { type: ["string", "null"], format: date-time }
        totalCost: { type: number }
        rationale: { type: ["string", "null"] }
        triggerNote: { type: ["string", "null"] }
        triggerSource: { type: ["string", "null"], description: "ai | manual | system" }
        createdAt: { type: string, format: date-time }
        createdByRole: { type: ["string", "null"] }
        sourceRecommendationId: { type: ["string", "null"] }
...
    Recommendation:
      type: object
      required: [id, kind, nodeId, itemId, quantity, priority, rationale, suggestedSupplierId, etaDays, generatedAt]
      properties:
        id: { type: string }
        kind: { type: string }
        nodeId: { type: string }
        nodeName: { type: string }
        itemId: { type: string }
        itemName: { type: string }
        quantity: { type: number }
        priority: { type: string }
        rationale: { type: string }
        suggestedSupplierId: { type: ["string", "null"] }
        suggestedSupplierName: { type: ["string", "null"] }
        suggestedFromNodeId: { type: ["string", "null"] }
        sourceChannel:
          type: ["string", "null"]
          description: Procurement channel family of the suggested supplier.
          enum: [DOD, COMMERCIAL, HOST_NATION, ALLIED, null]
        etaDays: { type: number }
        estimatedCost: { type: number }
        estimatedUnitCostUsd: { type: number }
        estimatedTotalCostUsd: { type: number }
        alternatives:
          type: array
          description: Up to 4 ranked alternative suppliers across channels.
          items: { $ref: "#/components/schemas/RecommendationAlternative" }
        displacedFrom:
          oneOf:
            - $ref: "#/components/schemas/RecommendationDisplacement"
            - type: "null"
          description: |
            When the scenario knocked the would-be primary supplier offline or
            below capacity and the COA had to reroute through an alternate,
            this records the displaced supplier and its degraded availability.
        splitAllocation:
          oneOf:
            - type: array
              items: { $ref: "#/components/schemas/RecommendationSplitAllocation" }
            - type: "null"
          description: |
            Capacity-aware split sourcing across two suppliers. Present only
            when the chosen primary's availability is constrained and a
            healthier alternate exists; the planner divides the total qty
            proportional to capacity.
        generatedAt: { type: string, format: date-time }
        confidenceScore: { type: number }
        scenarioId: { type: ["string", "null"] }
        promotedOrderId: { type: ["string", "null"] }
        companionItems:
          type: array
          description: |
            Primary-tier supply dependents drawn from the procedure(s) that
            consume this item. Used by the Recommendations rail to show a
            "+N companion supplies" chip and by the Promote dialog to bundle
            the dependents into one multi-line order.
          items: { $ref: "#/components/schemas/CompanionItem" }
    CompanionItem:
      type: object
      required: [itemId, itemName, tier, quantityPerEvent]
      properties:
        itemId: { type: string }
        itemName: { type: string }
        tier:
          type: string
          enum: [primary, secondary, tertiary]
        quantityPerEvent: { type: number }
        unit: { type: ["string", "null"] }
        onHandAtNode: { type: ["number", "null"] }
        procedureId: { type: ["string", "null"] }
        procedureName: { type: ["string", "null"] }
...
    RecommendationAlternative:
      type: object
      required: [supplierId, supplierName, channel, etaDays, reliabilityScore, estimatedUnitCostUsd, estimatedTotalCostUsd, rankScore]
      properties:
        supplierId: { type: string }
        supplierName: { type: string }
        channel:
          type: string
          enum: [DOD, COMMERCIAL, HOST_NATION, ALLIED]
        etaDays: { type: number }
        reliabilityScore: { type: number }
        estimatedUnitCostUsd: { type: number }
        estimatedTotalCostUsd: { type: number }
        rankScore: { type: number }
    RecommendationDisplacement:
      type: object
      required: [supplierId, supplierName, availabilityFraction]
      properties:
        supplierId: { type: string }
        supplierName: { type: string }
        availabilityFraction:
          type: number
          description: Capacity multiplier the scenario applied (0..1).
        cause: { type: ["string", "null"] }
    RecommendationSplitAllocation:
      type: object
      required: [supplierId, supplierName, channel, qty, pctOfTotal, etaDays]
      properties:
        supplierId: { type: string }
        supplierName: { type: string }
        channel:
          type: string
          enum: [DOD, COMMERCIAL, HOST_NATION, ALLIED]
        qty: { type: number }
        pctOfTotal:
          type: number
          description: Fraction of the total qty assigned to this supplier (0..1).
        etaDays: { type: number }
    PromoteRecommendationOverrides:
      type: object
      properties:
        quantity: { type: number, minimum: 1 }
        supplierId: { type: string }
        etaDays: { type: number, minimum: 0 }
        priority: { type: string, enum: [ROUTINE, URGENT, FLASH] }
        includeCompanionSupplies:
          type: boolean
          description: |
            When true, the promoted purchase order also includes one line per
            companion supply item (items that share a procedure tier with the
            recommended item). Each line uses the companion's recommended
            quantity and a covering supplier. Defaults to false.
    ForecastInput:
      type: object
      required: [itemIds, nodeIds, horizonDays]
      properties:
        itemIds:
          type: array
          items: { type: string }
        nodeIds:
          type: array
          items: { type: string }
        horizonDays: { type: integer }
        operationalState: { type: ["string", "null"] }
    ForecastResult:
      type: object
      required: [series]
      properties:
        series:
          type: array
          items:
            type: object
            required: [nodeId, itemId, points]
            properties:
              nodeId: { type: string }
              itemId: { type: string }
              itemName: { type: string }
              points:
                type: array
                items: { $ref: "#/components/schemas/HistoryPoint" }
    HistoryPoint:
      type: object
      required: [day, value]
      properties:
        day: { type: integer }
        value: { type: number }
        label: { type: ["string", "null"] }
...