# lib/api-zod/src/generated/types/historyPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * INDOPACOM IRONVEIN (Resilient Operational Network for Vital Expeditionary Inventory Nodes) API for medical logistics decision support.
 * OpenAPI spec version: 1.0.0
 */

export interface HistoryPoint {
  day: number;
  value: number;
  /** @nullable */
  label?: string | null;
}